update rename app (#6)
app renamed
