+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.myapp",
+    package = "seni.tembe.orbyt",
 )
 
 android_resource(
     name = "res",
-    package = "com.myapp",
+    package = "seni.tembe.orbyt",
     res = "src/main/res",
 )
