+++ android/app/src/main/java/seni/tembe/orbyt/MainApplication.java
@@ -1,4 +1,4 @@
-package com.myapp;
+package seni.tembe.orbyt;
 
 import android.app.Application;
 import android.content.Context;
@@ -62,7 +62,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.myapp.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("seni.tembe.orbyt.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
