+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "MyApp",
-  "displayName": "MyApp"
+  "name": "orbyt",
+  "displayName": "orbyt"
 }
