+++ ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0'
 
-target 'MyApp' do
+target 'orbyt' do
   config = use_native_modules!
 
   use_react_native!(
@@ -12,7 +12,7 @@ target 'MyApp' do
     :hermes_enabled => false
   )
 
-  target 'MyAppTests' do
+  target 'orbytTests' do
     inherit! :complete
     # Pods for testing
   end
